+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0b2 - unreleased
 ------------------
 
+* Handle case where inspect.getmembers hits an AttributeError.
+  [davisagli]
+
 * Handle case where ZODB state is not a mapping.
   [davisagli]