+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 1.0b2 - unreleased
 ------------------
 
+* Handle case where objects providing interfaces not present in the Python
+  environment can't be unpickled even into Broken objects.
+  [davisagli]
+
+* Cast non-string dict keys into strings for the sake of traversal.
+  [davisagli]
+
 * Handle case where inspect.getmembers hits an AttributeError.
   [davisagli]